+++ packages/surveys/src/utils/results-availability/index.ts
@@ -1,4 +1,3 @@
 export * from "./has-user-responded";
 export * from "./migrate-form-properties-settings";
 export * from "./should-display-results";
-export * from "./types";
