+++ packages/surveys/test/mocks/form-item-published.ts
@@ -1,7 +1,7 @@
 /* Copyright (c) 2018 Environmental Systems Research Institute, Inc.
  * Apache-2.0 */
 
-import { IFormItem } from "../../src/utils/results-availability";
+import { IFormItem } from "../../src/types";
 
 export const FormItemPublished: IFormItem = {
   id: "a5db32e043f14f6a9edfec7075288df6",
